complete exercise 01-cascade-fix
Completed the first and only cascade exercise utilizing numerous of the
rules we learned. I modified the ordering of the para/small-para
definitions to take advantage of last definition for matching
specificity. Used an ID selector for strongest specificity of the
confirm button. Lastly, used a class class chain which is more specific
than a type class selector.